+++ 07.bowling_object/lib/shot.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class Shot
+  attr_reader :mark
+
+  def initialize(mark)
+    @mark = mark
+  end
+
+  def score
+    mark.include?('X') ? 10 : mark.to_i
+  end
+end
+
