+++ 07.bowling_object/test/bowling_test.rb
@@ -41,12 +41,7 @@ class ShotTest < Minitest::Test
     game_score4 = Game.new('6390038273X9180XX00')
     assert_equal 134, game_score4.score
 
-    game_score4 = Game.new('XXXXXXXXXXXX')
-    assert_equal 300, game_score4.score
-  end
-
-  def test_hoge
-    p game = Game.new('6390038273X9180XX00')
-    p game.score
+    game_score5 = Game.new('XXXXXXXXXXXX')
+    assert_equal 300, game_score5.score
   end
 end
